// src/resources/Update_userUserInterface.ts
export enum Update_userUserInterface {

    buttom_account = '//*[@id="top-links"]/ul/li[2]/a',
    buttom_my_account = '//*[@id="top-links"]/ul/li[2]/ul/li[1]/a',
    buttom_edit_account = '//*[@id="content"]/ul[1]/li[1]/a',
    input_first_name = '//*[@id="input-firstname"]',
    input_last_name = '//*[@id="input-lastname"]',
    input_e_mail = '//*[@id="input-email"]',
    input_telephone = '//*[@id="input-telephone"]',
    buttom_continue = '//*[@id="content"]/form/div/div[2]/input',

}